requirements.txt
fastapi
uvicorn
openai
spacy
python-dotenv 